[kb] Move AzureOpenAIEncoder to its own file
To conform with our coding style
