+++ src/canopy/knowledge_base/record_encoder/__init__.py
@@ -1,3 +1,4 @@
 from .base import RecordEncoder
 from .dense import DenseRecordEncoder
-from .openai import OpenAIRecordEncoder, AzureOpenAIRecordEncoder
+from .openai import OpenAIRecordEncoder
+from .azure_openai import AzureOpenAIRecordEncoder
