+++ src/canopy/knowledge_base/record_encoder/openai.py
@@ -1,5 +1,5 @@
 from typing import List
-from pinecone_text.dense.openai_encoder import OpenAIEncoder, AzureOpenAIEncoder
+from pinecone_text.dense.openai_encoder import OpenAIEncoder
 from canopy.knowledge_base.models import KBDocChunk, KBEncodedDocChunk, KBQuery
 from canopy.knowledge_base.record_encoder.dense import DenseRecordEncoder
 from canopy.models.data_models import Query
@@ -53,29 +53,3 @@ class OpenAIRecordEncoder(DenseRecordEncoder):
         raise NotImplementedError
 
 
-class AzureOpenAIRecordEncoder(OpenAIRecordEncoder):
-    """
-    AzureOpenAIRecordEncoder is a type of DenseRecordEncoder that uses the OpenAI `embeddings` API.
-    The implementation uses the `AzureOpenAIEncoder` class from the `pinecone-text` library.
-    For more information about see: https://github.com/pinecone-io/pinecone-text
-
-    """  # noqa: E501
-
-    def __init__(
-        self,
-        *,
-        model_name: str = "text-embedding-ada-002",
-        batch_size: int = 400,
-        **kwargs
-    ):
-        """
-        Initialize the AzureOpenAIRecordEncoder
-
-        Args:
-            model_name: The name of the OpenAI embeddings model to use for encoding. See https://platform.openai.com/docs/models/embeddings
-            batch_size: The number of documents or queries to encode at once.
-                        Defaults to 400.
-            **kwargs: Additional arguments to pass to the underlying `pinecone-text. OpenAIEncoder`.
-        """  # noqa: E501
-        encoder = AzureOpenAIEncoder(model_name, **kwargs)
-        super().__init__(dense_encoder=encoder, batch_size=batch_size)
